feat(trading): size and account for fractional shares (Stage 4)
Stage 4 of the money-representation plan. Removes the whole-units guard so the
account replay accepts fractional fills, and lets the live equity path size
fractional shares.

- domain/accounting: drop _require_whole_units; apply_buy/apply_sell accept
  fractional quantities. Exact Decimal (live) / float (backtest) arithmetic over
  truncated quantities keeps a fully-sold position at exactly zero, so the
  _compact_positions qty > 0 test stays reliable.
- sizing: choose_buy_qty / allocate_buy_quantities / closing_sell_qty take a
  quantity_step and return fractional quantities. WHOLE_SHARE_STEP (the default)
  preserves the backtest and options/leaps at whole units; the live equity path
  passes FRACTIONAL_SHARE_STEP (the storage precision).
- selection: equity buys and sells size fractional; leaps stay whole.
- risk gate: BookTradeCandidate.qty and RiskGateDecision quantities are float; a
  buy within limits passes through at its exact fractional size, and a rescaled
  buy floors to a whole unit (conservative and instrument-agnostic).
- check_cash_invariant: exact integer reconciliation (no float tolerance), since
  money is stored as integer minor units.

+++ src/backtesting/services/simulation.py
@@ -238,7 +238,9 @@ def _execute_buys(
     # Every buy is sized against the same post-sell equity, so sizing does not
     # drift as earlier buys in the list execute.
     post_sell_equity = state.cash + compute_market_value(state.positions, trade_prices.to_dict())
-    sized_buys: list[tuple[str, float, int]] = []
+    # The backtest sizes whole shares: choose_buy_qty/allocate_buy_quantities default
+    # to WHOLE_SHARE_STEP, so these float quantities are whole-valued.
+    sized_buys: list[tuple[str, float, float]] = []
     for ticker in tickers:
         if signals[ticker] != "buy" or ticker not in active_tickers or state.positions[ticker] > 0:
             continue
